Code Edit

+++ src/main/java/com/jeonguk/web/service/impl/PostServiceImpl.java
@@ -1,44 +1,46 @@
 package com.jeonguk.web.service.impl;
 
-import java.util.concurrent.TimeUnit;
-
-import org.springframework.beans.factory.annotation.Autowired;
+import com.jeonguk.web.entity.Post;
+import com.jeonguk.web.exception.ResourceNotFoundException;
+import com.jeonguk.web.repository.PostRepository;
+import com.jeonguk.web.service.PostService;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
 
-import com.jeonguk.web.entity.Post;
-import com.jeonguk.web.repository.PostRepository;
-import com.jeonguk.web.service.PostService;
-
-import lombok.extern.slf4j.Slf4j;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 
 @Slf4j
+@AllArgsConstructor
 @Service
 public class PostServiceImpl implements PostService {
 
-    @Autowired
-    private PostRepository postRepository;
-
-    @Autowired
-    private RedisTemplate<String, Post> redisTemplate;
+    private final PostRepository postRepository;
+    private final RedisTemplate<String, Post> redisTemplate;
 
     @Override
     public Post findPostById(Long id) {
-        String key = "post_" + id;
-        ValueOperations<String, Post> operations = redisTemplate.opsForValue();
-        boolean hasKey = redisTemplate.hasKey(key);
+        final String key = "post_" + id;
+        final ValueOperations<String, Post> operations = redisTemplate.opsForValue();
+        final boolean hasKey = redisTemplate.hasKey(key);
         if (hasKey) {
-            Post post = operations.get(key);
+            final Post post = operations.get(key);
             log.info("PostServiceImpl.findPostById() : cache post >> " + post.toString());
             return post;
         }
-        Post post = postRepository.findOne(id);
-        operations.set(key, post, 10, TimeUnit.SECONDS);
-        log.info("PostServiceImpl.findPostById() : cache insert >> " + post.toString());
-        return post;
+        final Optional<Post> post = Optional.ofNullable(postRepository.findOne(id));
+        if(post.isPresent()) {
+            operations.set(key, post.get(), 10, TimeUnit.SECONDS);
+            log.info("PostServiceImpl.findPostById() : cache insert >> " + post.get().toString());
+            return post.get();
+        } else {
+            throw new ResourceNotFoundException();
+        }
     }
 
     public Page<Post> getAllPosts(Integer page, Integer size) {
@@ -52,8 +54,8 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public Post updatePost(Post post) {
-        String key = "post_" + post.getId();
-        boolean hasKey = redisTemplate.hasKey(key);
+        final String key = "post_" + post.getId();
+        final boolean hasKey = redisTemplate.hasKey(key);
         if (hasKey) {
             redisTemplate.delete(key);
             log.info("PostServiceImpl.updatePost() : cache delete >> " + post.toString());
@@ -63,13 +65,18 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public void deletePost(Long id) {
-        String key = "post_" + id;
-        boolean hasKey = redisTemplate.hasKey(key);
+        final String key = "post_" + id;
+        final boolean hasKey = redisTemplate.hasKey(key);
         if (hasKey) {
             redisTemplate.delete(key);
             log.info("PostServiceImpl.deletePost() : cache delete ID >> " + id);
         }
-        postRepository.delete(id);
+        final Optional<Post> post = Optional.ofNullable(postRepository.findOne(id));
+        if(post.isPresent()) {
+            postRepository.delete(id);
+        } else {
+            throw new ResourceNotFoundException();
+        }
     }
     
 }
